Added League table calculations, added api to retrieve all fixtures with predicted scorelines

+++ src/main/kotlin/com/cshep4/wcpredictor/data/Prediction.kt
@@ -4,5 +4,5 @@ data class Prediction(val id: Long = 0,
         var hGoals: Int? = null,
         var aGoals: Int? = null,
         var userId: Int? = null,
-        var matchId: Int? = null
+        var matchId: Long? = null
 )
