+++ src/main/kotlin/com/cshep4/wcpredictor/entity/PredictionEntity.kt
@@ -12,7 +12,7 @@ data class PredictionEntity (
         var hGoals: Int? = null,
         var aGoals: Int? = null,
         var userId: Int? = null,
-        var matchId: Int? = null
+        var matchId: Long? = null
 ){
     fun toDto(): Prediction = Prediction(
             id = this.id,
